+++ babel.config.js
@@ -6,11 +6,11 @@ module.exports = {
       {
         bugfixes: true,
         targets: {
-          esmodules: true,
-        },
-      },
+          esmodules: true
+        }
+      }
     ],
     '@babel/preset-react',
-    '@babel/preset-typescript',
-  ],
+    '@babel/preset-typescript'
+  ]
 }