+++ demo/src/sandboxes/gesture-drag/src/App.jsx
@@ -22,11 +22,10 @@ function Draggable() {
     boundToParent: false,
     rubberband: false,
     threshold: { value: [3, 3], step: 10 },
-    delay: { value: 1000, step: 100, min: 0, max: 3000, optional: true, disabled: true },
-    keys: true
+    delay: { value: 1000, step: 100, min: 0, max: 3000, optional: true, disabled: true }
   })
 
-  const pointerOptions = useControls('pointer', { touch: false, capture: true, lock: false, mouse: false })
+  const pointerOptions = useControls('pointer', { touch: false, capture: true, lock: false, mouse: false, keys: true })
 
   const bind = useDrag(
     ({ active, tap, ...state }) => {