+++ packages/core/src/engines/PinchEngine.ts
@@ -2,6 +2,7 @@ import { Engine } from './Engine'
 import { touchDistanceAngle, distanceAngle, wheelValues } from '../utils/events'
 import { V } from '../utils/maths'
 import { Vector2, WebKitGestureEvent } from '../types'
+import { clampStateInternalMovementToBounds } from '../utils/state'
 
 const SCALE_ANGLE_RATIO_INTENT_DEG = 30
 const PINCH_WHEEL_RATIO = 100
@@ -265,6 +266,9 @@ export class PinchEngine extends Engine<'pinch'> {
     state._delta = [(-wheelValues(event)[1] / PINCH_WHEEL_RATIO) * state.offset[0], 0]
     V.addTo(state._movement, state._delta)
 
+    // _movement rolls back to when it passed the bounds.
+    clampStateInternalMovementToBounds(state)
+
     this.state.origin = [event.clientX, event.clientY]
 
     this.compute(event)
@@ -292,6 +296,8 @@ export class PinchEngine extends Engine<'pinch'> {
     }
     // we try to set a passive listener, knowing that in any case React will
     // ignore it.
-    bindFunction('wheel', '', this.wheel.bind(this), { passive: false })
+    if (this.config.pinchOnWheel) {
+      bindFunction('wheel', '', this.wheel.bind(this), { passive: false })
+    }
   }
 }