+++ packages/core/src/engines/WheelEngine.ts
@@ -1,6 +1,7 @@
 import { CoordinatesEngine } from './CoordinatesEngine'
 import { wheelValues } from '../utils/events'
 import { V } from '../utils/maths'
+import { clampStateInternalMovementToBounds } from '../utils/state'
 
 export interface WheelEngine extends CoordinatesEngine<'wheel'> {
   wheel(this: WheelEngine, event: WheelEvent): void
@@ -23,17 +24,7 @@ export class WheelEngine extends CoordinatesEngine<'wheel'> {
     V.addTo(state._movement, state._delta)
 
     // _movement rolls back to when it passed the bounds.
-    const [ox, oy] = state.overflow
-    const [dx, dy] = state._delta
-    const [dirx, diry] = state._direction
-
-    if ((ox < 0 && dx > 0 && dirx < 0) || (ox > 0 && dx < 0 && dirx > 0)) {
-      state._movement[0] = state._movementBound[0] as number
-    }
-
-    if ((oy < 0 && dy > 0 && diry < 0) || (oy > 0 && dy < 0 && diry > 0)) {
-      state._movement[1] = state._movementBound[1] as number
-    }
+    clampStateInternalMovementToBounds(state)
 
     this.compute(event)
     this.emit()