+++ packages/core/src/types/internalConfig.ts
@@ -46,6 +46,7 @@ export type InternalDragOptions = Omit<InternalCoordinatesOptions<'drag'>, 'axis
   }
   delay: number
   axisThreshold: Record<PointerType, number>
+  keyboardDisplacement: number
 }
 
 export type InternalPinchOptions = InternalGestureOptions<'pinch'> & {
@@ -55,6 +56,7 @@ export type InternalPinchOptions = InternalGestureOptions<'pinch'> & {
   device: 'gesture' | 'pointer' | 'touch' | undefined
   lockDirection: boolean
   modifierKey: ModifierKey
+  pinchOnWheel: boolean
 }
 
 type MoveAndHoverMouseOnly = {