+++ packages/core/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @use-gesture/core
 
+## 10.2.24
+
+### Patch Changes
+
+- 60aae2149: feat: Added option to configure keyboard displacement in the drag gesture
+
+## 10.2.23
+
+### Patch Changes
+
+- 79684a05f: types: add package exports
+
+## 10.2.22
+
+### Patch Changes
+
+- c6215e8ad: fix: properly resolve pointer.keys config
+
+## 10.2.21
+
+### Patch Changes
+
+- 6f4c09b55: fix: rolls back wheel-based pinch movement to bounds (thanks [@Andarist](https://github.com/Andarist)!)
+- 854f4dfc1: feat (pinch):
+
+  - `pinchOnwheel: false` prevents pinching with wheel.
+
+## 10.2.20
+
+### Patch Changes
+
+- de807fddc: fix: applying a new config to useGesture / Gesture shouldn't throw an error.
+
 ## 10.2.19
 
 ### Patch Changes